refactor: move controller into v1alpha specific controller
