+++ controllers/v1alpha1/helper_types.go
@@ -1,4 +1,4 @@
-package controllers
+package v1alpha1
 
 // LagoonEnvironmentVariable is used to define Lagoon environment variables.
 type LagoonEnvironmentVariable struct {
