+++ controllers/v1alpha1/helpers_test.go
@@ -1,4 +1,4 @@
-package controllers
+package v1alpha1
 
 import (
 	"testing"
