+++ controllers/v1alpha1/lagoon_harborintegration.go
@@ -1,4 +1,4 @@
-package controllers
+package v1alpha1
 
 import (
 	"fmt"
