+++ controllers/v1alpha1/lagoonmonitor_controller.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package v1alpha1
 
 import (
 	"bytes"
