+++ controllers/v1alpha1/lagoonmonitor_taskhandlers.go
@@ -1,4 +1,4 @@
-package controllers
+package v1alpha1
 
 // this file is used by the `lagoonmonitor` controller
